Create Delete without head pointer
You have a singly linked list of size n  and given a pointer to a node in the list that needs to be removed. Any information about the head pointer or any other node is not given. The function should take only one argument, which is a pointer to the node to be deleted.
Note: No head reference is given to you. It is guaranteed that the node to be deleted is not a tail node in the linked list.

Example 1:

Input:
n = 2
value[] = {1,2}
node = 1
Output: 2
Explanation: After deleting 1 from the
linked list, we have remaining nodes
as 2.
Example 2:

Input:
n = 4
value[] = {10,20,4,30}
node = 20
Output: 10 4 30
Explanation: After deleting 20 from
the linked list, we have remaining
nodes as 10, 4, 30.
Your Task:
You only need to complete the function deleteNode that takes reference to the node that needs to be deleted. The function should take only one argument, which is a pointer to the node to be deleted. The printing is done automatically by the driver code.

Expected Time Complexity : O(1).
Expected Auxilliary Space : O(1).

Constraints:
2 <= n <= 103  
1 <= value[i] <= 109
1 <= node <= n-1